refactor: move DocViewer comp to seperate file
