+++ src/DocViewer.tsx
@@ -1,0 +1,60 @@
+import React, { CSSProperties, FC, memo } from "react";
+import styled, { ThemeProvider } from "styled-components";
+import { HeaderBar } from "./components/HeaderBar";
+import { ProxyRenderer } from "./components/ProxyRenderer";
+import { defaultTheme } from "./defaultTheme";
+import { AvailableLanguages } from "./i18n";
+import { DocRenderer, IConfig, IDocument, ITheme } from "./models";
+import { DocViewerRenderers } from "./renderers";
+import { DocViewerProvider } from "./store/DocViewerProvider";
+
+export interface DocViewerProps {
+  documents: IDocument[];
+  className?: string;
+  style?: CSSProperties;
+  config?: IConfig;
+  theme?: ITheme;
+  pluginRenderers?: DocRenderer[];
+  prefetchMethod?: string;
+  requestHeaders?: Record<string, string>;
+  initialActiveDocument?: IDocument;
+  language?: AvailableLanguages;
+  activeDocument?: IDocument;
+  onDocumentChange?: (document: IDocument) => void;
+}
+
+const DocViewer: FC<DocViewerProps> = (props) => {
+  const { documents, theme } = props;
+
+  if (!documents) {
+    throw new Error("Please provide an array of documents to DocViewer!");
+  }
+
+  return (
+    <DocViewerProvider pluginRenderers={DocViewerRenderers} {...props}>
+      <ThemeProvider
+        theme={theme ? { ...defaultTheme, ...theme } : defaultTheme}
+      >
+        <Container
+          id="react-doc-viewer"
+          data-testid="react-doc-viewer"
+          {...props}
+        >
+          <HeaderBar />
+          <ProxyRenderer />
+        </Container>
+      </ThemeProvider>
+    </DocViewerProvider>
+  );
+};
+
+export default memo(DocViewer);
+
+const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+  background: #ffffff;
+  width: 100%;
+  height: 100%;
+`;
